# docs/api/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
...
paths:
  /api/v1/events:
    get:
      tags:
      - Events
      summary: Return all active events
      description: Retrieve all active events
      operationId: getAllEvents
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Successfully retrieved the events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponseDTO"
    put:
      tags:
      - Events
      summary: Edit an Event
      description: Allows the admins to edit the prices of the sectors within a given
        event
      operationId: editEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditEventRequest"
        required: true
      responses:
        "400":
          description: The number of sectors selected for editing does not match the
            number of new prices
          content:
            application/json: {}
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Event or Sector Not Found
          content:
            application/json: {}
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Successfully modified the event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponseDTO"
    post:
      tags:
      - Events
      summary: Creates an Event
      description: Allows the admins to create an event
      operationId: createEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventCreateRequest"
        required: true
      responses:
        "400":
          description: The number of sectors selected for the creation process does
            not match the number of new prices
          content:
            application/json: {}
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Match Not Found
          content:
            application/json: {}
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "201":
          description: Successfully created the event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponseDTO"
...
  /api/v1/tickets/{ticketId}:
    get:
      tags:
      - Tickets
      summary: Get a certain ticket
      description: Return the ticket whose ID is provided
      operationId: getTicket
      parameters:
      - name: ticketId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Ticket Not Found
          content:
            application/json: {}
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Successfully retrieved the ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDTO"
...
  /api/v1/events/{eventId}:
    get:
      tags:
      - Events
      summary: Return a certain event
      description: Return the event whose ID is provided
      operationId: getEvent
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Event Not Found
          content:
            application/json: {}
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Successfully retrieved the event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponseDTO"
    delete:
      tags:
      - Events
      summary: Delete an Event
      description: Allows the admins to delete the given event
      operationId: deleteEvent
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Event Not Found
          content:
            application/json: {}
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Successfully deleted the event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponseDTO"
  /api/v1/events/{eventId}/tickets:
    get:
      tags:
      - Events
      summary: Get all sold tickets of an event
      description: Retrieves all sold tickets of a certain event
      operationId: getTicketsOfEvent
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Event Not Found
          content:
            application/json: {}
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Successfully retrieved the tickets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDTO"
  /api/v1/events/{eventId}/sectors:
    get:
      tags:
      - Events
      summary: Return available sectors of an event
      description: Retrieve all the available sectors of the given event
      operationId: getAvailableSectors
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Event Not Found
          content:
            application/json: {}
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Successfully retrieved the sectors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventManagerDTO"
  /api/v1/events/{eventId}/sector/{sectorId}:
    get:
      tags:
      - Events
      summary: Return available seats of an event
      description: Retrieve all the available seats of a given sector and event
      operationId: getAvailableSeats
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: sectorId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Event or Sector Not Found
          content:
            application/json: {}
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Successfully retrieved the seats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeatDTO"
  /api/v1/events/match/{matchId}:
    get:
      tags:
      - Events
      summary: Return the event of a certain match
      description: Returns the event belonging to the match whose ID is provided
      operationId: getEventByMatchId
      parameters:
      - name: matchId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Match Not Found
          content:
            application/json: {}
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Successfully retrieved the event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponseDTO"
...
components:
  schemas:
    EditEventRequest:
      type: object
      properties:
        eventId:
          type: integer
          format: int64
        sectorIds:
          type: array
          items:
            type: integer
            format: int64
          maxItems: 2147483647
          minItems: 1
        prices:
          type: array
          items:
            type: number
            format: double
          maxItems: 2147483647
          minItems: 1
      required:
      - eventId
      - prices
      - sectorIds
    EventManagerDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        sectorId:
          type: integer
          format: int64
        sectorName:
          type: string
        price:
          type: number
          format: double
        availability:
          type: integer
          format: int32
        totalCapacity:
          type: integer
          format: int32
    EventResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        awayTeamName:
          type: string
        homeTeamName:
          type: string
        homeTeamAbbreviation:
          type: string
        stadiumName:
          type: string
        date:
          type: string
          format: date-time
        pictureMap:
          $ref: "#/components/schemas/PictureInfo"
        sectors:
          type: array
          items:
            $ref: "#/components/schemas/EventManagerDTO"
    PictureInfo:
      type: object
      properties:
        url:
          type: string
        publicId:
          type: string
...
    SeatDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    TicketDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        awayTeamName:
          type: string
        homeTeamName:
          type: string
        stadiumName:
          type: string
        price:
          type: number
          format: double
        matchDate:
          type: string
          format: date-time
        sectorName:
          type: string
        seatName:
          type: string
...
    EventCreateRequest:
      type: object
      properties:
        matchId:
          type: integer
          format: int64
        prices:
          type: array
          items:
            type: number
            format: double
        sectors:
          type: array
          items:
            $ref: "#/components/schemas/SectorCreateRequest"
          maxItems: 2147483647
          minItems: 1
      required:
      - matchId
      - prices
      - sectors
    SectorCreateRequest:
      type: object
      properties:
        name:
          type: string
        totalCapacity:
          type: integer
          format: int32
...